requirements.txt
apparmor==2.13.3
appdirs==1.4.3
awscli==1.17.5
botocore==1.14.5
Brlapi==0.7.0
btrfsutil==1.1.1
CacheControl==0.12.6
cachetools==4.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
docutils==0.14
Flask==1.1.1
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
html5lib==1.0.1
httplib2==0.17.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
lensfun==0.3.2
LibAppArmor==2.13.3
louis==3.12.0
MarkupSafe==1.1.1
more-itertools==8.1.0
msgpack==0.6.2
oauthlib==3.1.0
ordered-set==3.1.1
packaging==20.0
pep517==0.6.0
progress==1.5
psycopg2-binary==2.8.4
pwquality==1.4.2
py3status==3.24
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.18.2
PyGObject==3.34.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytoml==0.1.21
pyxdg==0.26
PyYAML==5.3
rados==2.0.0
rbd==2.0.0
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rgw==2.0.0
rsa==4.0
s3transfer==0.3.1
six==1.14.0
team==1.0
uritemplate==3.0.1
urllib3==1.25.8
virtualenv==16.7.9
webencodings==0.5.1
Werkzeug==0.16.1
zipp==0.6.0
